2024.09.02 피드백 관련 코드 파일생성 및 코드 리팩토링
