+++ src/constants/champion.ts
@@ -1,0 +1,7 @@
+export type CHAMPION = {
+  key: number;
+  id: string;
+  name: string;
+  image_url: string;
+  position: string[];
+};
